rename Map component to not use index.js
